# docs/api/openapi.supplement.yaml
openapi: 3.0.3
info:
  title: Hearth API — supplement
  description: |
    Hand-written routes that have no proto service definition yet.
    These are merged with docs/api/openapi.proto-derived.json by `make openapi`
    to produce docs/api/openapi.json (the authoritative merged spec).
    When HEA-969 adds proto definitions for these endpoints, remove them here.
  version: 0.1.0
paths:
...
  /oauth/consents:
    get:
      summary: List the authenticated user's active consents
      operationId: selfListConsents
      tags: [OAuth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of active consents
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT